# sms-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "School Management System — Go + Gin + PostgreSQL\n\n## Authentication\nThis API uses JWT Bearer tokens. Login via `POST /api/login` to receive an access token (1 hour) and a refresh token (7 days).\nClick the **Authorize** button above and enter: `Bearer \u003cyour_token\u003e`\n\n## Roles\n| Role    | Access                                              |\n|---------|-----------------------------------------------------|\n| Admin   | Full access to all endpoints                        |\n| Teacher | Attendance, grades, locker (read public)             |\n| Student | Own grades, report card, locker, finance             |\n| Parent  | Children's data, report cards, finance               |",
        "title": "SMS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/notify/absences": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Looks up all absences for the given date (or today if omitted) and sends an email to each student's parent. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Notify parents of absent students",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date to check absences for (YYYY-MM-DD). Defaults to today.",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications sent",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid date format",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Admin only",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/notify/broadcast": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends an email announcement to all active users, or to a specific set of roles. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Broadcast an announcement",
                "parameters": [
                    {
                        "description": "Announcement data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.BroadcastInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Announcement sent",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or invalid role",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Admin only",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/locker/files/{fileID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Students can only delete their own files. Admins can delete any file. Removes from disk and database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locker"
                ],
                "summary": "Delete a locker file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "File ID",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File deleted",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Not your file",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/locker/files/{fileID}/visibility": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Toggles a locker file between private (teacher cannot see) and public (teacher can see). Student must own the file.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locker"
                ],
                "summary": "Toggle file visibility",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "File ID",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visibility updated",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Not your file",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/locker/student/{studentID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Students see all their own files. Teachers see only public files. Admins see all files.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locker"
                ],
                "summary": "List files in a student's locker",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "studentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File list",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Cannot view another student's files",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Student profile not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/locker/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Students can upload files (max 5MB) of allowed types: pdf, jpg, jpeg, png, doc, docx, txt. Each file is stored under the student's locker directory.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locker"
                ],
                "summary": "Upload a file to the digital locker",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload (max 5MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category: Certificate | Assignment | Portfolio | Material",
                        "name": "category",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "File uploaded",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "File missing, too large, or invalid type",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — Students only",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Student profile not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.BroadcastInput": {
            "type": "object",
            "required": [
                "body",
                "title"
            ],
            "properties": {
                "body": {
                    "type": "string",
                    "minLength": 10,
                    "example": "Please collect report cards by Friday."
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Student\"",
                        "\"Parent\"]"
                    ]
                },
                "title": {
                    "type": "string",
                    "minLength": 5,
                    "example": "End of Term Reminder"
                }
            }
        },
...
        "helpers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT token. Example: **Bearer eyJhbGci...**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}